tests/test_libs_dataloader.py
import pytest
from unittest.mock import Mock, patch, MagicMock
import tempfile
import os
from src.libs.dataloader import Loader
from src.libs.dataloader.document import DocumentLoader
from src.libs.dataloader.web import WebLoader
from src.core.rag.schema import Document, DocumentChunk, Source


class TestLoader:
    @pytest.fixture
    def temp_path(self):
        with tempfile.TemporaryDirectory() as temp_dir:
            yield temp_dir

    def test_init_success(self, temp_path):
        loader = Loader(temp_path)
        assert loader.path == temp_path
        assert loader.recursive is True
        assert loader.chunk_size == 1024
        assert loader.chunk_overlap == 200

    def test_init_with_custom_params(self, temp_path):
        loader = Loader(temp_path, recursive=False, chunk_size=512, chunk_overlap=100)
        assert loader.path == temp_path
        assert loader.recursive is False
        assert loader.chunk_size == 512
        assert loader.chunk_overlap == 100

    def test_init_file_not_found(self):
        with pytest.raises(ValueError, match="File not found: /nonexistent"):
            Loader("/nonexistent")

    def test_create_source_not_implemented(self, temp_path):
        loader = Loader(temp_path)
        with pytest.raises(NotImplementedError, match="Subclasses should implement this method"):
            loader.create_source()

    def test_load_data_not_implemented(self, temp_path):
        loader = Loader(temp_path)
        with pytest.raises(NotImplementedError, match="Subclasses should implement this method"):
            loader.load_data()


class TestDocumentLoader:
    @pytest.fixture
    def temp_path(self):
        with tempfile.TemporaryDirectory() as temp_dir:
            yield temp_dir

    @pytest.fixture
    def temp_file(self):
        with tempfile.NamedTemporaryFile(mode='w', suffix='.txt', delete=False) as f:
            f.write("Test content for document loader")
            temp_file_path = f.name
        yield temp_file_path
        os.unlink(temp_file_path)

    def test_init_success(self, temp_path):
        loader = DocumentLoader(temp_path)
        assert loader.path == temp_path
        assert hasattr(loader, 'sentence_splitter')

    def test_create_source_success(self, temp_path):
        loader = DocumentLoader(temp_path)
        source_path = "/test/file.txt"
        
        source = loader.create_source(source_path)
        
        # Check if it has Source-like attributes instead of isinstance check
        assert hasattr(source, 'name')
        assert hasattr(source, 'type')
        assert hasattr(source, 'uri')
        assert hasattr(source, 'id')
        assert source.name == "file.txt"
        assert source.type == "file"
        assert source.uri == source_path

    def test_create_source_empty_path(self, temp_path):
        loader = DocumentLoader(temp_path)
        loader.path = ""  # Set empty path
        
        with pytest.raises(ValueError, match="Path cannot be empty"):
            loader.create_source("/test/file.txt")

    @patch('src.libs.dataloader.document.SimpleDirectoryReader')
    @patch('llama_index.core.node_parser.SentenceSplitter.get_nodes_from_documents')
    def test_load_data_success(self, mock_splitter_method, mock_reader, temp_path):
        loader = DocumentLoader(temp_path)
        
        # Mock the SimpleDirectoryReader
        mock_doc = Mock()
        mock_doc.text = "Test document content"
        mock_doc.metadata = {"file_path": "/test/file.txt"}
        
        mock_reader_instance = Mock()
        mock_reader_instance.load_data.return_value = [mock_doc]
        mock_reader.return_value = mock_reader_instance
        
        # Mock sentence splitter method
        mock_node = Mock()
        mock_node.text = "Test chunk content"
        mock_splitter_method.return_value = [mock_node]

        # Execute the generator
        results = list(loader.load_data())
        
        assert len(results) == 1
        source, document, chunks = results[0]
        
        # Check attributes instead of isinstance for Source
        assert hasattr(source, 'name')
        assert hasattr(source, 'type')
        assert hasattr(document, 'content')
        assert hasattr(document, 'title')
        assert isinstance(chunks, list)
        assert len(chunks) == 1
        assert hasattr(chunks[0], 'content')

    @patch('src.libs.dataloader.document.SimpleDirectoryReader')
    def test_load_data_no_documents(self, mock_reader, temp_path):
        loader = DocumentLoader(temp_path)
        
        mock_reader_instance = Mock()
        mock_reader_instance.load_data.return_value = []
        mock_reader.return_value = mock_reader_instance
        
        with pytest.raises(ValueError, match="Failed to load document"):
            list(loader.load_data())

    @patch('src.libs.dataloader.document.SimpleDirectoryReader')
    def test_load_data_empty_document(self, mock_reader, temp_path):
        loader = DocumentLoader(temp_path)
        
        # Mock document with empty text
        mock_doc = Mock()
        mock_doc.text = ""
        mock_doc.metadata = {"file_path": "/test/file.txt"}
        
        mock_reader_instance = Mock()
        mock_reader_instance.load_data.return_value = [mock_doc]
        mock_reader.return_value = mock_reader_instance
        
        # Should skip empty documents
        results = list(loader.load_data())
        assert len(results) == 0

    @patch('src.libs.dataloader.document.SimpleDirectoryReader')
    def test_load_data_exception(self, mock_reader, temp_path):
        loader = DocumentLoader(temp_path)
        
        mock_reader.side_effect = Exception("Reader error")
        
        with pytest.raises(ValueError, match="Failed to load document"):
            list(loader.load_data())


class TestWebLoader:
    @pytest.fixture
    def web_loader(self):
        return WebLoader("https://example.com")

    def test_init_success(self):
        loader = WebLoader("https://example.com", url_pattern=".*", max_urls=5000, chunk_size=512, chunk_overlap=100)
        
        assert loader.url == "https://example.com"
        assert loader.path == "https://example.com"  # For compatibility
        assert loader.max_urls == 5000
        assert loader.chunk_size == 512
        assert loader.chunk_overlap == 100
        assert hasattr(loader, 'url_pattern')
        assert hasattr(loader, 'visited_urls')
        assert hasattr(loader, 'found_urls')
        assert hasattr(loader, 'sentence_splitter')

    def test_create_source_success(self, web_loader):
        source_url = "https://example.com/page"
        
        source = web_loader.create_source(source_url)
        
        # Check if it has Source-like attributes instead of isinstance check
        assert hasattr(source, 'name')
        assert hasattr(source, 'type') 
        assert hasattr(source, 'uri')
        assert hasattr(source, 'id')
        assert source.name == "example.com"
        assert source.type == "website"
        assert source.uri == source_url

    @patch('src.libs.dataloader.web.requests.get')
    def test_visit_site_success(self, mock_get, web_loader):
        mock_response = Mock()
        mock_response.text = "<html><body><h1>Test</h1><p>Content</p></body></html>"
        mock_response.raise_for_status.return_value = None
        mock_get.return_value = mock_response
        
        # Mock _get_urls to return empty list
        web_loader._get_urls = Mock(return_value=[])
        
        content, links = web_loader._visit_site("https://example.com")
        
        assert "Test" in content
        assert "Content" in content
        assert isinstance(links, list)
        mock_get.assert_called_once()

    @patch('src.libs.dataloader.web.requests.get')
    def test_visit_site_retry_on_failure(self, mock_get, web_loader):
        mock_get.side_effect = Exception("Connection error")
        
        with pytest.raises(ValueError, match="Failed to fetch content"):
            web_loader._visit_site("https://example.com", retry=1)

    @patch('src.libs.dataloader.web.requests.get')
    def test_get_urls_success(self, mock_get, web_loader):
        html_content = '''
        <html>
            <body>
                <a href="/page1">Page 1</a>
                <a href="https://example.com/page2">Page 2</a>
                <a href="#fragment">Fragment</a>
            </body>
        </html>
        '''
        
        mock_response = Mock()
        mock_response.text = html_content
        mock_response.raise_for_status.return_value = None
        mock_get.return_value = mock_response
        
        from bs4 import BeautifulSoup
        soup = BeautifulSoup(html_content, 'html.parser')
        
        # Since _get_urls calls requests.get again, we need to mock it
        urls = web_loader._get_urls(soup, "https://example.com")
        
        assert isinstance(urls, list)

    @patch('llama_index.core.node_parser.SentenceSplitter.get_nodes_from_documents')
    def test_load_data_basic_structure(self, mock_splitter_method, web_loader):
        # Mock all the methods that load_data uses
        web_loader._visit_site = Mock(return_value=("Test content", []))
        
        # Create a real Source object for testing
        from src.core.rag.schema import Source
        test_source = Source(name="test", type="website", uri="https://example.com")
        web_loader.create_source = Mock(return_value=test_source)
        
        # Mock sentence splitter method
        mock_node = Mock()
        mock_node.text = "Test chunk"
        mock_splitter_method.return_value = [mock_node]

        results = list(web_loader.load_data())
        
        # Should have at least one result if there's content
        if results:
            source, document, chunks = results[0]
            assert hasattr(source, 'name')
            assert hasattr(source, 'type')
            assert hasattr(document, 'content')
            assert hasattr(document, 'title')
            assert isinstance(chunks, list)

    def test_load_data_skip_empty_content(self, web_loader):
        # Mock _visit_site to return empty content
        web_loader._visit_site = Mock(return_value=("", []))
        
        # Create a real Source object for testing
        from src.core.rag.schema import Source
        test_source = Source(name="test", type="website", uri="https://example.com")
        web_loader.create_source = Mock(return_value=test_source)
        
        results = list(web_loader.load_data())
        
        # Should skip empty content
        assert len(results) == 0

    def test_load_data_exception_handling(self, web_loader):
        # Mock _visit_site to raise an exception
        web_loader._visit_site = Mock(side_effect=Exception("Network error"))
        
        with pytest.raises(ValueError, match="Failed to process URL"):
            list(web_loader.load_data())
