requirements.txt
-e .
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.9
attrs==25.3.0
bandit==1.8.3
banks==2.1.2
beautifulsoup4==4.13.4
black==25.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coverage==7.8.0
cryptography==44.0.3
dataclasses-json==0.6.7
Deprecated==1.2.18
deprecation==2.1.0
dill==0.4.0
dirtyjson==1.0.8
distro==1.9.0
docker==7.1.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
flake8==7.2.0
frozenlist==1.6.0
fsspec==2025.3.2
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.2
griffe==1.7.3
h11==0.16.0
hf-xet==1.1.0
html2text==2025.4.15
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
llama-cloud==0.1.19
llama-cloud-services==0.6.22
llama-index==0.12.35
llama-index-agent-openai==0.4.7
llama-index-cli==0.4.1
llama-index-core==0.12.35
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-openai==0.3.38
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.22
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
mcp==1.8.0
mdurl==0.1.2
mock==5.2.0
mpmath==1.3.0
multidict==6.4.3
mypy==1.15.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
openai==1.78.1
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pbr==6.1.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.5.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
pylint==3.3.7
pymgclient==1.4.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
pypdf==5.5.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.3.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sse-starlette==2.3.4
starlette==0.46.2
stevedore==5.4.1
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
testcontainers==3.7.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
wrapt==1.17.2
yarl==1.20.0
